addd new table txs

+++ alonaBatrak/new table txs.sql
@@ -1,7 +1,7 @@
  create table txs(
 id int primary key auto_increment,
 created_at timestamp default current_timestamp,
-amount int,
+amount float,
 is_approved boolean,
 sender_account_id int not null,
 recipient_account_id int not null
